style.css: 	+ Adjusted the color of a link after being clicked
extractnodes_syscheck-for-listMatching.ksh:
	+ Minor optimization of script

gen-webcontent.ksh:
	+ Named daily.local.ksh to gen-webcontent.ksh

hostInfoReports.ksh:
	+ Optimized the process of producing header

listMatching.ksh:
	+ Changed error and output redirection

listMatchingReports.ksh:
	+ Changed error and output redirection
	+ Cut space at the beginning of the output number for
	Yes_NA_Report.txt

vx.ksh:
	+ Added full path of commands

+++ extractnodes_syscheck-for-listMatching.ksh
@@ -28,7 +28,7 @@
 	#
 	#
 	# CHANGELOG:
-	# Mar 28 2016 PHAT TRAN
+	# Apr 1 2016 PHAT TRAN
 	############################################################################################################
 
 	HOST=`uname -n | cut -d'.' -f1`
@@ -53,7 +53,4 @@
 	done
 	
 	export size=`du -h $TARGETDIR/syscheck-$HOST.list | awk '{print $1}'`
-	if [  "$size" == "0K" ]
-	then 
-		rm -rf $TARGETDIR
-	fi
+	if [  "$size" == "0K" ] && rm -rf $TARGETDIR
